// models/AccountInfoData.ts
// src/models/AccountInfoData.ts

export interface AccountInfoData {
  title: string;
  password: string;
  days: string;
  months: string;
  years: string;
  newsletter: boolean;
  optin: boolean;
  firstName: string;
  lastName: string;
  company: string;
  address1: string;
  country: string;
  state: string;
  city: string;
  zipcode: string;
  mobileNumber: string;
}